Return long instead of int when we are counting categories.
Should be in 54c9ab9a899c1afbc5f7021f2c8db55a34b6a27d commit.

No functional changes.

+++ src/main/java/ru/mystamps/web/dao/CategoryDao.java
@@ -25,8 +25,8 @@ import ru.mystamps.web.entity.Category;
 import ru.mystamps.web.service.dto.SelectEntityDto;
 
 public interface CategoryDao extends PagingAndSortingRepository<Category, Integer> {
-	int countByName(String name);
-	int countByNameRu(String name);
+	long countByName(String name);
+	long countByNameRu(String name);
 	
 	@Query(
 		"SELECT NEW ru.mystamps.web.service.dto.SelectEntityDto("
